Bagging, which stands for Bootstrap Aggregating, is an ensemble learning technique designed to improve the stability and accuracy of machine learning algorithms.
Bagging, which stands for Bootstrap Aggregating, is an ensemble learning technique designed to improve the stability and accuracy of machine learning algorithms.

It works particularly well with high-variance models like decision trees. The primary idea behind bagging is to combine the predictions of multiple models trained on different subsets of the data to reduce variance and prevent overfitting.